+++ src/layout/LineSeparator.ts
@@ -1,0 +1,6 @@
+import styled from "../styled-components";
+
+export const LineSeparator = styled.hr`
+    border: none;
+    border-top: 1px dotted ${({theme}) => theme.colors.separator};
+`;